Addressed PR feedback
- TODO for another PR, change Kotlin generator to use emptyMap() instead of mapOf() where appropriate
